Converted `File` to Rust.
The next step is to add (de)serialization.
Since invalid rules are simply `io::Error`, the `rules` array has been
converted into a `HashMap` which also includes the name of the rule,
such that even if a rule is invalid a suitable error message can be
returned to the user (`Rule 'foo' is invalid!` is better than
`No rule 'foo'`).