Redone error handling
There were two major changes:
1. The `custom_error` crate is now used for `prelude::Error`.
2. The `File::rules` type has changed to remove the `smake::Result`
   layer (to become `HashMap<String, Rule>`). The file parser will
   collect any errors from rules and return them at construction time
   itself.

The first part makes the prelude code much, much simpler (though the
syntax is a little bit weird).

The second part completely changes how stuff works. Now, if the parsed
SMakefile is missing an input (and a help message is not being
displayed), the user is immediately given an error about the missing
input, even if they are not using the target which requires the input.