Restarting with a spec
The code's starting to look clunky and undignified, and now I realise
that it's because I have no plan in mind. So I've written up a spec to
work with, which includes all kinds of information about the SMake I'm
trying to create. Note that the spec is not complete, and it will be
updated as time goes on. I may move it to the wiki if I change it often.

I have also disabled the test that was causing Travis CI to fail. It was
due to the fact that the `File::from_file()` function will now return an
`Err` if the input file does not exist, which was not the case
previously.